refact: unify animations for FullScreenDialog
