+++ apps/store/src/components/ChangeSsnWarningDialog/ChangeSsnWarningDialog.tsx
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled'
-import { motion, type Transition } from 'framer-motion'
 import { useState, useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Button, Text, WarningTriangleIcon, theme } from 'ui'
@@ -36,39 +35,23 @@ export const ChangeSsnWarningDialog = ({ open, onAccept, onDecline }: Props) => 
       <FullscreenDialog.Modal
         center={true}
         Footer={
-          <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            transition={ANIMATE_TRANSITION}
-          >
+          <>
             <Button loading={loading} onClick={handleAccept}>
               {t('CHANGE_SSN_BUTTON')}
             </Button>
             <Button variant="ghost" onClick={onDecline}>
               {t('DIALOG_BUTTON_CANCEL', { ns: 'common' })}
             </Button>
-          </motion.div>
+          </>
         }
       >
-        <motion.div
-          initial={{
-            opacity: 0,
-            y: '20%',
-          }}
-          animate={{
-            opacity: 1,
-            y: 0,
-          }}
-          transition={ANIMATE_TRANSITION}
-        >
-          <IconWithText>
-            <WarningTriangleIcon size="1em" color={theme.colors.amber600} />
-            <Text align="center">{t('CHANGE_SSN_TITLE')}</Text>
-          </IconWithText>
-          <Text color="textSecondary" align="center">
-            {t('CHANGE_SSN_DESCRIPTION')}
-          </Text>
-        </motion.div>
+        <IconWithText>
+          <WarningTriangleIcon size="1em" color={theme.colors.amber600} />
+          <Text align="center">{t('CHANGE_SSN_TITLE')}</Text>
+        </IconWithText>
+        <Text color="textSecondary" align="center">
+          {t('CHANGE_SSN_DESCRIPTION')}
+        </Text>
       </FullscreenDialog.Modal>
     </FullscreenDialog.Root>
   )
@@ -80,9 +63,3 @@ const IconWithText = styled(Text)({
   justifyContent: 'center',
   alignItems: 'center',
 })
-
-const ANIMATE_TRANSITION: Transition = {
-  duration: 0.6,
-  delay: 0.3,
-  ...theme.transitions.framer.easeInOutCubic,
-}
